refactor(snap): add sub command run to fix auto completion
This enables the swift argument parser to correctly generate the
auto-completion scripts so they are actually usable.
